requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.12.5
cfgv==3.3.0
distlib==0.3.1
Django==3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
gunicorn==20.1.0
identify==2.2.4
nodeenv==1.6.0
pre-commit==2.12.1
psycopg2==2.8.6
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
wincertstore==0.2
